Move some values to `.env*` files
Values such as API keys, services urls, etc. Maybe could be more appropriate put
them in a configuration file instead as a environment variables, but for
simplicity reasons I prefer to follow the Hanami "conventions".

This commit also performs changes in the Weather an Country interactors,
allowing them to receive a Hash as single param.